+++ R/get-set-convenience.R
@@ -286,6 +286,8 @@ set_cell_properties <- function (ht, row, col, ...) {
 #' ```
 #' @return The `ht` object.
 #'
+#' @seealso merge_repeated_rows
+#'
 #' @export
 #' @examples
 #' ht <- hux(a = 1:3, b = 1:3)
@@ -318,7 +320,52 @@ merge_cells <- function (ht, row, col) {
 }
 
 
-get_all_borders <- function(ht, row, col, drop = TRUE) {
+#' Merge repeated rows into multirow cells
+#'
+#' @param ht A huxtable.
+#' @param row A row specification.
+#' @param col A column specification.
+#'
+#' @return The modified huxtable.
+#' @export
+#'
+#' @details
+#' Repeated rows in each column are merged into cells with
+#' `rowspan > 1`.
+#'
+#' If `row` contains gaps, results may be unexpected (and a warning is given).
+#'
+#' @seealso merge_cells
+#'
+#' @examples
+#' ht <- as_hux(jams[c(1, 2, 2, 3, 3, 4), ])
+#' ht <- add_columns(ht, c("Sugar", "30%", "40%", "30%", "40%", "30%"),
+#'       after = 1)
+#' ht
+#' merge_repeated_rows(ht)
+#' merge_repeated_rows(ht, everywhere, "Type")
+merge_repeated_rows <- function (ht, row, col) {
+  assert_that(is_huxtable(ht))
+  row <- get_rc_spec(ht, row, 1)
+  col <- get_rc_spec(ht, col, 2)
+  if (is.logical(row)) row <- which(row)
+  if (is.logical(col)) col <- which(col)
+
+  if (length(row) > 1L && ! all(row == seq(min(row), max(row)))) {
+    warning("Non-contiguous rows: ", paste(row, collapse = ", "))
+  }
+  for (cc in col) {
+    contents <- ht[row, ][[cc]] # gets a vector
+    new <- which(c(TRUE, contents[seq_len(length(contents) - 1)] != contents[-1]))
+    spans <- diff(c(new, length(contents) + 1))
+    rowspan(ht)[row[new], cc] <- spans
+  }
+
+  ht
+}
+
+
+get_all_borders <- function (ht, row, col, drop = TRUE) {
   list(
     left   = left_border(ht)[row, col, drop = drop],
     right  = right_border(ht)[row, col, drop = drop],
@@ -328,7 +375,7 @@ get_all_borders <- function(ht, row, col, drop = TRUE) {
 }
 
 
-get_all_border_colors <- function(ht, row, col, drop = TRUE) {
+get_all_border_colors <- function (ht, row, col, drop = TRUE) {
   list(
     left   = left_border_color(ht)[row, col, drop = drop],
     right  = right_border_color(ht)[row, col, drop = drop],
@@ -338,7 +385,7 @@ get_all_border_colors <- function(ht, row, col, drop = TRUE) {
 }
 
 
-get_all_border_styles <- function(ht, row, col, drop = TRUE) {
+get_all_border_styles <- function (ht, row, col, drop = TRUE) {
   list(
     left   = left_border_style(ht)[row, col, drop = drop],
     right  = right_border_style(ht)[row, col, drop = drop],
@@ -348,7 +395,7 @@ get_all_border_styles <- function(ht, row, col, drop = TRUE) {
 }
 
 
-get_all_padding <- function(ht, row, col, drop = TRUE) {
+get_all_padding <- function (ht, row, col, drop = TRUE) {
   list(
     left   = left_padding(ht)[row, col, drop = drop],
     right  = right_padding(ht)[row, col, drop = drop],