+++ NEWS.md
@@ -1,6 +1,19 @@
 
+
+
 Note that huxtable attempts to follow semantic versioning (https://semver.org). Therefore, major 
 version increments reflect backwards-incompatible API changes, not necessarily big changes.
+
+# huxtable 4.7.0
+
+* Better error messages.
+* New `merge_repeated_rows()` function: merge repeated rows into a single
+  cell.
+* New `fill` and `colspan/rowspan` arguments for `insert_row()/insert_column()`: 
+  - `insert_row(ht, "blah", "", "", "", "", ...)` can be written
+    `insert_row(ht, "blah", fill = "")`.
+  - `colspan/rowspan` set `colspan/rowspan` of the first cell in the inserted
+    row/column.
 
 # huxtable 4.6.1